requirements.txt
django==3.0.5
djoser==2.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
